sf crop name refactored
